requirements.txt
kafka-python
requests
pymongo
Faker